[REFACTOR] 도메인 패키지 구조로 변경
