--- secondSeminar/src/main/java/com/example/secondSeminar/dto/request/MemberProfileUpdateRequest.java
@@ -1,7 +1,0 @@
-package com.example.secondSeminar.dto.request;
-
-import com.example.secondSeminar.enums.Part;
-
-public record MemberProfileUpdateRequest(int generation, Part part) {
-}
-
