+++ secondSeminar/src/main/java/com/example/secondSeminar/member/dto/request/MemberCreateRequest.java
@@ -1,0 +1,7 @@
+package com.example.secondSeminar.member.dto.request;
+
+import com.example.secondSeminar.member.domain.SOPT;
+
+public record MemberCreateRequest(String name, String nickname, int age, SOPT sopt) {
+}
+
