+++ secondSeminar/src/main/java/com/example/secondSeminar/member/dto/request/MemberProfileUpdateRequest.java
@@ -1,0 +1,7 @@
+package com.example.secondSeminar.member.dto.request;
+
+import com.example.secondSeminar.common.enums.Part;
+
+public record MemberProfileUpdateRequest(int generation, Part part) {
+}
+
